Rename desstinationsRepo to destinations-repo
