+++ test/destinations-repo-test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import DestinationsRepo from '../src/destinationsRepo'
+import DestinationsRepo from '../src/destinations-repo'
 import TestData from '../src/test-data'
 
 describe('DestinationsRepo', () => {
